Updated 2D painting in the AsoBrain 3D model:  - Overloaded 'Node3D.paint()' method to allow paint of object using their own rendering hints (do not require application to always provide these).     > Added rendering hint properties to Object3D: outlineColor, fillColor, shadeFactor, alternateOutlineColor, and alternateFillColor.  - Optimizations:     > Inlined 'getAdjustedFillColor()' to prevent unnecessarily repeated calculations.     > Extracted 'Face3D.paint()' method from 'Object3D.paint()' method to paint a face.     > Changed 'shadeFactor' argument type of the 'Node3D.paint()' method from 'double' to 'float' (Java color model support only float precision).

+++ java/ab/j3d/model/Cylinder3D.java
@@ -214,7 +214,7 @@ public final class Cylinder3D
 		}
 	}
 
-	public void paint( final Graphics2D g , final Matrix3D gTransform , final Matrix3D viewTransform , final Color outlineColor , final Color fillColor , final double shadeFactor )
+	public void paint( final Graphics2D g , final Matrix3D gTransform , final Matrix3D viewTransform , final Color outlineColor , final Color fillColor , final float shadeFactor )
 	{
 		final Matrix3D viewBase = xform.multiply( viewTransform );
 		final double   h        = height;
